+++ projects/ngx-xchange-ui/src/components/card-metrics/card-metrics/card-metrics.component.ts
@@ -1,0 +1,25 @@
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'xc-card-metrics',
+  templateUrl: './card-metrics.component.html',
+  styleUrls: ['./card-metrics.component.css']
+})
+export class CardMetricsComponent {
+  @Input()
+    value: Number | null;
+  @Input()
+  description: string | null;
+  @Input()
+  width: string | null;
+
+  constructor() {
+    this.value = 0;
+    this.description = ""
+    this.width = '84'
+  }
+  
+  getWidth(){
+    return this.width!? this.width : '84'
+  }
+}